# lib/api-zod/src/generated/types/consensusAgentOutput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Model Orchestrator & Source Harvester API
 * OpenAPI spec version: 0.1.0
 */
import type { ConsensusAgentOutputEvidence } from "./consensusAgentOutputEvidence";
import type { ConsensusAgentOutputSignals } from "./consensusAgentOutputSignals";
import type { ConsensusAgentOutputVote } from "./consensusAgentOutputVote";

export interface ConsensusAgentOutput {
  agentId: string;
  agentName: string;
  vote: ConsensusAgentOutputVote;
  confidence: number;
  reasoning: string;
  latencyMs: number;
  evidence?: ConsensusAgentOutputEvidence;
  signals?: ConsensusAgentOutputSignals;
}