# lib/api-zod/src/generated/types/consensusGuardedAgent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Model Orchestrator & Source Harvester API
 * OpenAPI spec version: 0.1.0
 */
import type { ConsensusAgentOutput } from "./consensusAgentOutput";
import type { ConsensusGuardResult } from "./consensusGuardResult";

export interface ConsensusGuardedAgent {
  output: ConsensusAgentOutput;
  guard: ConsensusGuardResult;
}